Fixed pan camera, ints were not being converted to float.
git-svn-id: https://192.168.3.139:8443/svn/Engine/trunk@2049 146ae76c-2b19-ef4c-9a69-67eb2b888ad7

+++ Editor/Camera/OrbitCameraController.cs
@@ -94,8 +94,8 @@ namespace Editor
                 {
                     if (PanCamera.HeldDown)
                     {
-                        lookAt += rotatedLeft * (mouseCoords.x / (events.Mouse.AreaWidth * SCROLL_SCALE) * orbitDistance);
-                        lookAt += rotatedUp * (mouseCoords.y / (events.Mouse.AreaHeight * SCROLL_SCALE) * orbitDistance);
+                        lookAt += rotatedLeft * ((float)mouseCoords.x / (events.Mouse.AreaWidth * SCROLL_SCALE) * orbitDistance);
+                        lookAt += rotatedUp * ((float)mouseCoords.y / (events.Mouse.AreaHeight * SCROLL_SCALE) * orbitDistance);
                         updateTranslation(lookAt + normalDirection * orbitDistance);
                     }
                     else if (ZoomCamera.HeldDown)
